+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 #Change Log
 
+## Version 1.1.0
+* Support 3.x.x Android Plugin
+* Disable incremental mode
+
 ## Version 1.0.4
 *2018-01-25
 * Fix: in rare case, lancet generate illegal bytecode because of multi-thread.